feat: initial cleanup and cache fixes
