+++ src/lib/service/github.ts
@@ -1,8 +1,3 @@
-import type { SWRInfiniteResponse } from "swr/infinite";
-import useSWRInfinite from "swr/infinite";
-
-import { swrNoAutoUpdateSettings } from "./api";
-
 export interface Contributor {
   login: string;
   id: number;
@@ -12,19 +7,38 @@ export interface Contributor {
 
 const CONTRIBUTORS_BASE_URL = "https://api.github.com/repos/PaperMC/Paper/contributors?per_page=100";
 
-const fetcher = (url: string) => fetch(url).then((res) => res.json());
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const getURL = (pageIndex: number, previousPageData: any): string | null => {
-  if (previousPageData && previousPageData.length < 100) return null;
-  return `${CONTRIBUTORS_BASE_URL}&page=${pageIndex + 1}`;
-};
+export async function fetchContributors(maxPages = 15): Promise<Contributor[]> {
+  const all: Contributor[] = [];
 
-export const useGitHubContributors = (): SWRInfiniteResponse<Contributor[]> =>
-  useSWRInfinite(getURL, fetcher, swrNoAutoUpdateSettings);
+  for (let page = 1; page <= maxPages; page++) {
+    const res = await fetch(`${CONTRIBUTORS_BASE_URL}&page=${page}`, {
+      headers: {
+        Accept: "application/vnd.github+json",
+      },
+      next: { revalidate: 3600, tags: ["github:contributors"] },
+    });
+
+    if (res.status === 403) {
+      break;
+    }
+
+    if (!res.ok) {
+      break;
+    }
+
+    const data = (await res.json()) as Contributor[];
+    if (!Array.isArray(data) || data.length === 0) break;
+
+    all.push(...data);
+
+    if (data.length < 100) break;
+  }
+  return all;
+}
 
 export const getProjectRepository = (project: string, version: string): string => {
   if (project !== "paper") return `https://github.com/PaperMC/${project}`;
-  if (project == "paper" && version == "1.7.10") return "https://github.com/PaperMC/Paper-1.7";
+  if (project === "paper" && version === "1.7.10") return "https://github.com/PaperMC/Paper-1.7";
 
   const baseVersion = [21, 4]; // 1.21.4 is after the hardfork
   const isBelowBaseVersion = version
